+++ target_chains/solana/programs/core-bridge/src/legacy/state/guardian_set.rs
@@ -237,3 +237,64 @@ impl GuardianSet {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use {
+        super::LEGACY_GUARDIANS,
+        wormhole_sdk::{
+            core::{Action, GovernancePacket},
+            Vaa,
+        },
+        wormhole_solana::mainnet_guardian_sets::{
+            MAINNET_INITIAL_GUARDIAN, MAINNET_UPGRADE_GUARDIAN_SET_VAAS,
+        },
+    };
+
+    /// Wormhole testnet guardian (testnet guardian set 0). It never appears in the mainnet
+    /// upgrade VAAs but is part of `LEGACY_GUARDIANS` to cover testnet deployments.
+    const TESTNET_GUARDIAN: &str = "13947bd48b18e53fdaeee77f3473391ac727c638";
+
+    fn decode_guardian(hex_str: &str) -> [u8; 20] {
+        hex::decode(hex_str).unwrap().try_into().unwrap()
+    }
+
+    fn upgrade_vaa_guardians(vaa_hex: &str) -> (u32, Vec<[u8; 20]>) {
+        let vaa: Vaa<GovernancePacket> =
+            serde_wormhole::from_slice(&hex::decode(vaa_hex).unwrap()).unwrap();
+        match vaa.payload.action {
+            Action::GuardianSetUpgrade {
+                new_guardian_set_index,
+                new_guardian_set,
+            } => (
+                new_guardian_set_index,
+                new_guardian_set
+                    .addresses
+                    .into_iter()
+                    .map(|address| address.0)
+                    .collect(),
+            ),
+            action => panic!("VAA is not a guardian set upgrade: {action:?}"),
+        }
+    }
+
+    #[test]
+    fn test_legacy_guardians_match_upgrade_vaas() {
+        // `LEGACY_GUARDIANS` is expected to contain, in order of first appearance, every
+        // guardian that has ever been part of a mainnet guardian set (the initial set plus
+        // all upgrade VAAs), followed by the testnet guardian.
+        let mut expected: Vec<[u8; 20]> = vec![decode_guardian(MAINNET_INITIAL_GUARDIAN)];
+        for (i, vaa) in MAINNET_UPGRADE_GUARDIAN_SET_VAAS.iter().enumerate() {
+            let (new_guardian_set_index, guardians) = upgrade_vaa_guardians(vaa);
+            assert_eq!(new_guardian_set_index, i as u32 + 1);
+            for guardian in guardians {
+                if !expected.contains(&guardian) {
+                    expected.push(guardian);
+                }
+            }
+        }
+        expected.push(decode_guardian(TESTNET_GUARDIAN));
+
+        assert_eq!(LEGACY_GUARDIANS.to_vec(), expected);
+    }
+}
